Allow Linux tagging with spaces
When attempting to include tags with spaces, the tag was truncated after the first space. E.g.

```
-t GROUP "Tag with spaces"
```

Would result in a tag key value of `GROUP Tag`.

This is due to the way the `${tags[@]}"` expression handles a space delimted array as passed.

The fix gathers ALL arguments after the "key" and builds a quoted string to pass to the eventual command.